Update basics.py
Fixed Casting Error

+++ models/basics.py
@@ -32,7 +32,7 @@ def compl_mul1d(a, b):
     Yflip = flip_periodic(b)
     Yeven = 0.5 * (b + Yflip)
     Yodd  = 0.5 * (b - Yflip)
-    Z = torch.einsum("bix,iox->box", a, Yeven.long()) +  torch.einsum("bix,iox->box", Xflip, Yodd.long())
+    Z = torch.einsum("bix,iox->box", a, Yeven) +  torch.einsum("bix,iox->box", Xflip, Yodd)
     return Z    
 
 def compl_mul2d(a, b):
@@ -44,7 +44,7 @@ def compl_mul2d(a, b):
     Yflip = flip_periodic(b)
     Yeven = 0.5 * (b + Yflip)
     Yodd  = 0.5 * (b - Yflip)
-    Z =  torch.einsum("bixy,ioxy->boxy", a, Yeven.long()) + torch.einsum("bixy,ioxy->boxy", Xflip, Yeven.long())
+    Z =  torch.einsum("bixy,ioxy->boxy", a, Yeven) + torch.einsum("bixy,ioxy->boxy", Xflip, Yeven)
     return Z   
 
 def compl_mul3d(a, b):
@@ -56,7 +56,7 @@ def compl_mul3d(a, b):
     Yflip = flip_periodic(b)
     Yeven = 0.5 * (b + Yflip)
     Yodd  = 0.5 * (b - Yflip)
-    Z = torch.einsum("bixyz,ioxyz->boxyz", a, Yeven.long()) + torch.einsum("bixyz,ioxyz->boxyz", Xflip, Yodd.long())
+    Z = torch.einsum("bixyz,ioxyz->boxyz", a, Yeven) + torch.einsum("bixyz,ioxyz->boxyz", Xflip, Yodd)
     return Z   
 
 ################################################################
